file 6 ver2

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-income = int(input("Enter your monthly income: "))
-expences = int(input("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+expences = float(input("Enter your total monthly expenses: "))
 savings = income - expences
 interest_rate = 0.05
 total_saving = (savings * 12) + (savings * 12 * interest_rate)
